Servlet tests for JBoss-based adapters with TLS are broken
Fixes #20385
